+++ src/load_inputs/load_resources_data.jl
@@ -735,7 +735,8 @@ function validate_policy_dataframe!(filename::AbstractString, policy_in::DataFra
     accepted_cols = ["derating_factor", "esr", "esr_vrestor",
         "h2_demand", "qualified_supply",
         [string(cap, type) for cap in ["min_cap", "max_cap"]
-         for type in ("", "_stor", "_solar", "_wind")]...]
+         for type in ("", "_stor", "_solar", "_wind")]...,
+        "inv_incentive", "prod_incentive"]
 
     # Check that all policy columns have names in accepted_cols
     if !all(x -> replace(x, r"(_*|_*\d*)$" => "") in accepted_cols, cols)