+++ src/write_outputs/write_costs.jl
@@ -33,6 +33,12 @@ function write_costs(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
     if !isempty(ELECTROLYZER_ALL)
         push!(cost_list, "cHydrogenRevenue")
     end
+    if setup["InvestmentIncentive"] == 1
+        push!(cost_list, "cInvIncentiveBenefit")
+    end
+    if setup["ProductionIncentive"] == 1
+        push!(cost_list, "cProdIncentiveBenefit")
+    end
     dfCost = DataFrame(Costs = cost_list)
 
     cVar = value(EP[:eTotalCVarOut]) +
@@ -92,6 +98,14 @@ function write_costs(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
 
     if !isempty(ELECTROLYZER_ALL)
         push!(total_cost, -1 * value(EP[:eTotalHydrogenValue]))
+    end
+    
+    if setup["InvestmentIncentive"] == 1
+        push!(total_cost, -1 * value(EP[:eTotalInvIncentiveBenefit]))
+    end
+    
+    if setup["ProductionIncentive"] == 1
+        push!(total_cost, -1 * value(EP[:eTotalProdIncentiveBenefit]))
     end
 
     dfCost[!, Symbol("Total")] = total_cost